# optimize Generated{Abstract,Product}Algebra with benchmarking