… algebird-core/src/main/scala/com/twitter/algebird/GeneratedAbstractAlgebra.scala
… algebird-core/src/main/scala/com/twitter/algebird/GeneratedProductAlgebra.scala
… project/plugins.sbt
… scripts/ntuple_generators.rb
… scripts/product_generators.rb