This PR

- Adds a benchmark for `Tuple4` (seemed like a nice one to pick :)
- corrects some bad performance in the generated `sumOption` implementations that this benchmark uncovered.

Run the benchmark with:

```
sbt algebird-benchmark/jmh:run -t1 -f 1 -wi 5 -i 15 .*Tuple4Benchmark.*
```

### Benchmark Results

Before the changes:
```
[info] Benchmark                             (numElements)   Mode  Cnt     Score     Error  Units
[info] Tuple4Benchmark.timeProductPlus               10000  thrpt   15  1870.581 ± 149.159  ops/s
[info] Tuple4Benchmark.timeProductSumOption          10000  thrpt   15   492.128 ±  29.016  ops/s
[info] Tuple4Benchmark.timeTuplePlus                 10000  thrpt   15  3911.323 ± 106.611  ops/s
[info] Tuple4Benchmark.timeTupleSumOption            10000  thrpt   15  2682.713 ±  39.312  ops/s
```

After the changes:
```
[info] Benchmark                             (numElements)   Mode  Cnt     Score     Error  Units
[info] Tuple4Benchmark.timeProductPlus               10000  thrpt   15  1934.965 ±  50.165  ops/s
[info] Tuple4Benchmark.timeProductSumOption          10000  thrpt   15  2896.808 ±  71.128  ops/s
[info] Tuple4Benchmark.timeTuplePlus                 10000  thrpt   15  3893.420 ± 139.511  ops/s
[info] Tuple4Benchmark.timeTupleSumOption            10000  thrpt   15  4073.248 ±  60.992  ops/s
[success] Total time: 88 s, completed Dec 1, 2016 12:48:43 PM
```

so ~6x improvement for the ProductN semigroups, 1.5x improvement for the tupleN semigroups. AND, more importantly, they're all more efficient than just using `plus`.

Note that benchmarks only adds longs, so it measures the improvement of `sumOption` on tuples when the tuple elements don't have an efficient `sumOption` implementation. When they do, these `sumOption` implementations will be that much faster.