@sritchie 2016-12-01
@johnynek ah, nice catches. Was letting the coffee get to me. All fixed up.