@isnotinvain 2016-12-01
LGTM

Any idea why ignoring `size` didn't trigger a test failure? Is that just a size hint for performance?

Does this now always use sumOption, even when plus would normally be called? I wonder if any `plus(x,y)` implementations are more efficient than `sumOption(Seq(x,y))` in the case of only adding 2 items together? (eg, map monoid's sumOption creating a whole mutable hash map in sumOption)