@sritchie-stripe 2016-12-02
@isnotinvain, my mistake was that I left a value of `1000` hardcoded, instead of respecting the `size` parameter. The Tuple and Record `sumOption` implementations also pass a value of `1000`, so there was no difference.

All this PR does is add a more efficient implementation for anything that explicitly calls `sumOption` - the `plus` logic still works exactly like it did before. In fact, the benchmark calls out the difference between using `plus` explicitly vs `sumOption` for larger collections.

If you call `sumOption` on a collection of two things, it's true that it might be more efficient to just add them together with `plus` - `sumOption` operates on `TraversableOnce`, though, so it can't know that information. It's up to the caller to know the difference.

Thanks for the LGTM! Tests timed out, so waiting for the build to complete again and then I'll merge and continue with the averaged value implementation.