@johnynek 2016-12-02
so, I'm worried about this.

On a deep tuple are we exponentially growing buffers? Previously we were not (I don't think). You had a linear number of buffers. Can we stop and document why this is not exponential here?