@isnotinvain 2016-12-03
Sorry, looks like I pulled the trigger too soon. Did this go into the release? 
Do we need to back it out :( ?